requirements.txt
alembic==1.7.5
click==7.1.2
colorama==0.4.4
ecdsa==0.17.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==1.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
python-dateutil==2.8.1
python-editor==1.0.4
pyasn1==0.4.8
pycodestyle==2.8.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.28
typing-extensions==4.0.1
Werkzeug==1.0.1
zipp==3.6.0
python-dotenv
authlib==1.1.0
#authlib==0.15.2
#authlib==0.14.3
requests==2.24.0
